backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.34
psycopg2-binary==2.9.9
alembic==1.13.2
pydantic-settings==2.4.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt>=4.0.1,<5
python-dotenv==1.0.1
faiss-cpu
sentence-transformers
httpx==0.27.2
requests
langgraph
python-docx
PyPDF2
email-validator>=2.0.0